refactor: template builder
